requirements.txt
pygame>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0
